requirements.txt
alembic>=1.15.2
blinker>=1.9.0
click>=8.1.8
colorama>=0.4.6
Flask>=2.2.5  
Flask-Admin>=1.6.1
Flask-Login>=0.6.3
Flask-Migrate>=4.1.0
Flask-SQLAlchemy>=3.1.1
greenlet>=3.2.0
itsdangerous>=2.2.0
Jinja2>=3.1.6
Mako>=1.3.10
MarkupSafe>=3.0.2
python-dotenv>=0.21.1
SQLAlchemy>=2.0.40
typing_extensions>=4.13.2
Werkzeug>=3.1.3
WTForms>=3.2.1
psycopg2-binary>=2.9.0
requests>=2.31.0
oauthlib>=3.2.2
Flask-WTF>=0.15.1
flask-bootstrap>=3.3.7.1
flask-talisman>=1.1.0
flask-cors>=5.0.1
flask-limiter>=3.12
limits>=5.1.0
ordered-set>=4.1.0
rich>=13.9.4
deprecated>=1.2.18
markdown-it-py>=3.0.0
mdurl>=0.1.2
wrapt>=1.17.2
ollama>=0.4.8